Adding function to get the pressed key and handle/format it.
More info:
the function gets the pressed key and checks if it's a number,
if it's a number, add it to the end of the 'base_str' and remove the first (if the string only has zeros, and the number pressed is zero, nothing happens).
if the pressed key is not a number, then nothing happens.
at the moment, pressing 'ENTER' (return key) will only reset the string to '000',
pressing backspace will delete the last number if the string is bigger than 3, if not, it'll delete the last number and add a zero at the begining. Pressing it when there are only zeros in the string won't do anything