This part of the code is responsible for contolling the values in the 'pagamento_frame', the 'terminar' and 'cancelar' buttons state/color and if the 'dinheiro_faltando_label' should or should not be on the screen, when the return (ENTER) key is pressed.
More info:

if 'troco' is less than zero:
it will grid in the screen (only if isn't already there), bellow 'dinheiro_recebido_frame', a label to signal that the typed value is not enough.
if 'botao_terminar' is clickable in the screen, it will grey-it-out and turn it's lime color to the original color.
if the value in 'troco_do_dinheiro_frame' is diferent of "0.00" it will be changed to that.

else:
'troco_do_dinheiro_frame' value will be set to 'troco' value,
'botao_terminar' color will be changed to lime and it's state to
clickable,
'botao_cancelar' color will be changed to red.